+++ CHANGELOG.md
@@ -39,7 +39,7 @@ Main (unreleased)
 
 - Added additional backwards compatibility metrics to `prometheus.write.queue`. (@mattdurham)
 
-- Add new stdlib functions to_json (@ravishankar15)
+- Add new stdlib functions encoding.to_json (@ravishankar15)
 
 - Added OpenTelemetry logs and metrics support to Alloy mixin's dashboards and alerts. (@thampiotr)